+++ crates/codegen/src/statements/mod.rs
@@ -154,7 +154,7 @@ mod tests {
         let stmt = IrStmt::VarDecl(vars.clone());
         let tokens = stmt.codegen();
         let expected = quote::quote! {
-            let a: f64 = 1.0;
+            let a: f64 = (1) as f64;
             let mut b: bool = false;
         };
